+++ src/main/java/io/mosscloud/fibonaccinumbers/consumer/PrintConsumerStrategy.java
@@ -3,7 +3,7 @@ package io.mosscloud.fibonaccinumbers.consumer;
 import java.util.Collection;
 public class PrintConsumerStrategy implements FibonacciConsumerStrategy {
     @Override
-    public void consumerFibonacci(Collection<Integer> fibonacciCountResult) {
+    public void consumerFibonacci(Collection<Long> fibonacciCountResult) {
         System.out.println(fibonacciCountResult);
     }
 }
